turn postCreate schema into auth awared version - update routes in posts.py such that can authorize loggedin user

+++ routers/posts.py
@@ -9,6 +9,9 @@ import model
 from database import get_db
 from schemas import PostCreate, PostResponse, PostUpdate
 
+#import current_user
+from auth import CurrentUser
+
 router = APIRouter()
 
 
@@ -28,22 +31,16 @@ async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
 
 
 #Route to reponse with the CREATE method 
+#After adding current_user:CurrentUser, the route is protected with token authentication
 @router.post("", response_model=PostResponse, status_code=status.HTTP_201_CREATED)
-async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(
-        select(model.User).where(model.User.id == post.user_id),
-    )
-    user = result.scalars().first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
-
+async def create_post(
+    post: PostCreate, 
+    current_user:CurrentUser, 
+    db: Annotated[AsyncSession, Depends(get_db)]):
     new_post = model.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id,
+        user_id=current_user.id,
     )
     db.add(new_post)
     await db.commit()
@@ -74,7 +71,11 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
 #Route/endpoints to response to the UPDATE request for single posts
 #PUT --> full update
 @router.put("/{post_id}", response_model=PostResponse)
-async def update_post_full(post_id: int, post_data: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_full(
+    post_id: int, 
+    post_data: PostCreate,
+    current_user:CurrentUser, 
+    db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(
         select(model.Post)
         .options(selectinload(model.Post.author))
@@ -88,21 +89,28 @@ async def update_post_full(post_id: int, post_data: PostCreate, db: Annotated[As
 
     #Check if the client is reassigning the existing post's author,
     #verify the new data contain user_id exists before allowing the update
-    if post_data.user_id != post.user_id:
-        result = await db.execute(
-            select(model.User)
-            .where(model.User.id == post_data.user_id))
-        user = result.scalars().first()
-        if not user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="User not found",
-            )
+    #delete since we have added get_current_user dependecny
+    # if post_data.user_id != post.user_id:
+    #     result = await db.execute(
+    #         select(model.User)
+    #         .where(model.User.id == post_data.user_id))
+    #     user = result.scalars().first()
+    #     if not user:
+    #         raise HTTPException(
+    #             status_code=status.HTTP_404_NOT_FOUND,
+    #             detail="User not found",
+    #         )]
+
+    #add ownership track
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post"
+        )
 
     #update all the field in the post 
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
 
     #commit to the database after PUT opereation
     #no need to use db.add() because this is not insertion which require building new object
@@ -113,7 +121,11 @@ async def update_post_full(post_id: int, post_data: PostCreate, db: Annotated[As
 
 #PATCH --> partial update
 @router.patch("/{post_id}", response_model=PostResponse)
-async def update_post_partial(post_id: int, post_data: PostUpdate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_partial(
+    post_id: int, 
+    post_data: PostUpdate, 
+    current_user:CurrentUser,
+    db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(
         select(model.Post)
         .options(selectinload(model.Post.author))
@@ -125,6 +137,13 @@ async def update_post_partial(post_id: int, post_data: PostUpdate, db: Annotated
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post was not found")
 
     #No user_id in PostUpdate field --> remove user check 
+    
+    #add ownership check
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post"
+        )
 
     #Logic for PATCH --> only update fields that the client actually send 
     #post_data contain the data from the request body
@@ -146,7 +165,10 @@ async def update_post_partial(post_id: int, post_data: PostUpdate, db: Annotated
 
 #Route/endpoint to handle DELETE request
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(
+    post_id: int, 
+    current_user:CurrentUser,
+    db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(
         select(model.Post)
         .options(selectinload(model.Post.author))
@@ -155,6 +177,13 @@ async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]
     #check if the post exist to DELETE --> else 404 error
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post was not found")
+
+    #add ownership track
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to delete this post"
+        )
 
     await db.delete(post)
     await db.commit()
